flutter_chekmate/requirements.txt
selenium==4.15.2
webdriver-manager==4.0.1
Pillow==10.1.0

